Validate lobby ids without Joi, move MapImageInfo to common, and tidy the landing page
