+++ client/app-routes.tsx
@@ -76,11 +76,7 @@ export function AppRoutes({
       <Route path='/ladder/*?' component={LadderRouteComponent} />
       <Route path='/leagues/*?' component={LeagueRoot} />
       <Route path='/live' component={LiveStreamsPage} />
-      {IS_ELECTRON ? (
-        <Route path='/lobbies/:lobbyId/*?' component={LobbyView} />
-      ) : (
-        <Route path='/lobbies/:lobbyId/*?' component={LobbyLandingPage} />
-      )}
+      <Route path='/lobbies/:lobbyId/*?' component={IS_ELECTRON ? LobbyView : LobbyLandingPage} />
       <Route path='/maps/*?' component={MapsRoot} />
       <Route path='/news/:id/*?' component={NewsPostPage} />
       <Route path='/news' component={NewsArchivePage} />
