+++ client/gameplay-activity/gameplay-activity-widget.tsx
@@ -7,9 +7,9 @@ import styled, { css } from 'styled-components'
 import { Except } from 'type-fest'
 import { useRoute } from 'wouter'
 import { slotCount, takenSlotCount } from '../../common/lobbies'
+import { urlForLobby } from '../../common/lobbies/lobby-url'
 import { useObservedDimensions } from '../dom/dimension-hooks'
 import { MaterialIcon } from '../icons/material/material-icon'
-import { urlForLobby } from '../lobbies/lobby-url'
 import { cancelFindMatch } from '../matchmaking/action-creators'
 import { isInDraftAtom } from '../matchmaking/draft-atoms'
 import { ElapsedTime } from '../matchmaking/elapsed-time'
