+++ client/lobbies/lobby-url.ts
@@ -4,8 +4,6 @@ import { lobbySlug, urlForLobby } from '../../common/lobbies/lobby-url'
 import { SbLobbyId } from '../../common/lobbies/sb-lobby-id'
 import { push, replace } from '../navigation/routing'
 
-export { lobbySlug, urlForLobby }
-
 /** Navigates to a particular lobby. */
 export function navigateToLobby(id: SbLobbyId, name?: string, transitionFn = push): void {
   transitionFn(urlForLobby(id, name))
